@@ -22,8 +22,8 @@ use Innmind\Immutable\Str;
 
 final class SetupSite implements Command
 {
-    private $client;
-    private $config;
+    private Client $client;
+    private Adapter $config;
 
     public function __construct(Client $client, Adapter $config)
     {
